requirements.txt
Django==2.0.4
django-photologue==3.8.1
django-sortedm2m==1.5.0
djangorestframework==3.8.2
ExifRead==2.1.2
gunicorn==19.7.1
Pillow==5.1.0
psycopg2==2.7.4
pytz==2018.4
